// src/parser/generated/message-structure-schemas.ts
export const GENERATED_STRUCTURE_SCHEMAS: readonly PublishedStructureSchema[] = [
// ...
  {
// ...
      {
        name: "FINANCIAL_OBSERVATION_STANDALONE",
        kind: "group",
        position: 6,
        min: 0,
        max: "*",
        parent: 31,
      },
// ...
  },
// ...
];